requirements.txt
# Core UI
streamlit==1.44.1

# LLM APIs
openai==1.91.0
openai-agents>=0.0.8
agno==1.2.13
firecrawl-py
elevenlabs

# HTTP + Web
requests>=2.32.0
httpx>=0.24.0
duckduckgo-search>=6.3.7
google-search-results
newspaper4k>=0.9.3.1
lxml-html-clean>=0.4.1

# Data and analysis
pandas>=2.0.0
numpy==1.26.4
duckdb==1.1.3
openpyxl
yfinance>=0.2.0

# Code execution sandbox
e2b-code-interpreter==1.0.3
e2b==1.0.5

# Image + Audio processing
Pillow>=10.4.0
pillow==11.1.0
matplotlib>=3.7.0
plotly>=5.0.0
pydicom>=2.4.0

# Typing & Models
typing-extensions>=4.0.0
pydantic>=2.0.0

# Additional agents
phidata

# Built-in modules (DO NOT INSTALL via pip — shown here for clarity only):
# os, json, re, sys, io, uuid, asyncio, datetime, contextlib, base64, logging
